Restructured files and how they were divided
