+++ src/Pages/HomePage/Dashboards/CondoOwner/OwnerHome.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import { Link } from 'react-router-dom'; // Import Link for redirection
 import './OwnerHome.css'; // Import CSS file for styling
-import logo from '../../../imgs/Logos/Nexus-clear-noText.png';
+import logo from '../../../../Images/Logos/Nexus-clear-noText.png';
 
 
 
 //Import or fetch images for the condos
-import unit1 from "../../../imgs/CondoImages/Units/Unit1/img1.jpg";
-import unit2 from "../../../imgs/CondoImages/Units/Unit2/img1.jpg";
-import unit3 from "../../../imgs/CondoImages/Units/Unit3/img1.jpg";
-import unit4 from "../../../imgs/CondoImages/Units/Unit4/img1.jpg";
+import unit1 from "../../../../Images/CondoImages/Units/Unit1/img1.jpg";
+import unit2 from "../../../../Images/CondoImages/Units/Unit2/img1.jpg";
+import unit3 from "../../../../Images/CondoImages/Units/Unit3/img1.jpg";
+import unit4 from "../../../../Images/CondoImages/Units/Unit4/img1.jpg";
