+++ src/Pages/HomePage/Home.js
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
-import Logout from "../components/auth/Logout/Logout";
-import Center from "../components/utils/Center";
-import Profile from "./Profile/Profile";
+import Logout from "../../Components/Auth/Logout/Logout";
+import Center from "../../Components/Utils/Center";
+import Profile from "../ProfilePage/Profile";
 
 const Home = (props) => {
   useEffect(() => {}, []); // eslint-disable-line react-hooks/exhaustive-deps
