+++ src/Pages/ProfilePage/MyAccount/MyAccount.js
@@ -1,6 +1,6 @@
 import { Box, Button, TextField, Typography } from "@mui/material";
 import { useContext } from "react";
-import AuthenticatedUserContext from "../../../contexts/AuthenticatedUserContext";
+import AuthenticatedUserContext from '../../../Contexts/AuthenticatedUserContext';
 
 import "./MyAccount.css";
