+++ src/components/auth/AuthContainer/AuthContainer.js
@@ -1,11 +1,11 @@
 import { useState } from "react";
 import { signInWithEmailAndPassword, signInWithPopup } from "firebase/auth";
 import { useNavigate } from "react-router-dom";
-import { auth, Providers } from "../../config/firebase";
+import { auth, Providers } from "../../../config/firebase";
 import { Box, Button, TextField, Typography } from "@mui/material";
 import GoogleIcon from "@mui/icons-material/Google";
 import LoginIcon from '@mui/icons-material/Login';
-import Center from "../utils/Center";
+import Center from "../../Utils/Center";
 
 import "./AuthContainer.css";
